Add develop branch add Jacoco, Mockito, Junit 4 in pom.xml
