+++ src/main/java/com/praktikum/IngredientType.java
@@ -1,0 +1,11 @@
+package com.praktikum;
+
+/**
+ * Перечисление с типами ингредиентов.
+ * SAUCE – соус
+ * FILLING – начинка
+ */
+public enum IngredientType {
+    SAUCE,
+    FILLING
+}
