--- src/main/java/praktikum/Bun.java
@@ -1,25 +1,0 @@
-package praktikum;
-
-/**
- * Модель булочки для бургера.
- * Булочке можно дать название и назначить цену.
- */
-public class Bun {
-
-    public String name;
-    public float price;
-
-    public Bun(String name, float price) {
-        this.name = name;
-        this.price = price;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public float getPrice() {
-        return price;
-    }
-
-}
